QOL tweak: explain what "Invert button" means. (#1086)
I felt there was some uncertainty about the meaning of "invert". Given
the context it would seem to be most likely relating to an inversion of
the functionality (sneak vs no sneak), however given its proximity to
the extensive positioning options, it was unclear to me until I
triple-checked in the source code exactly what it did.

It probably doesn't matter, but at least this commit message is longer
even than the commit itself.