Fix considerable overhead caused by uncached reflection.
My first instinct was to actually replace this with a proper cached
reflection call, but given the number of times that this is potentially
being called on a server with a large number of players (to the point
where it is causing `EntityItem::onUpdate` to take up around 7% of tick
time), this seems like the situation that access transformers were
*explicitly designed for.

I feel as though even cached reflection would still add additional
overhead.

Especially given that this code is injected irregardless of the setting
being enabled (it being a client-side setting), speed would seem to be a
priority here.