Simplify BigDungeonChestProcessor loot table code.
This code creates an inherently worldless tile entity (which becomes
problematic with `fillWithLoot`), sets a null loot table (when the
default is already null), replaces the contents with empty item stacks
(when they should in theory already be null), and then sets the loot
table.

It then serializes the tile entity back to NBT and returns it.

Given that we have access to the tile entity NBT originally, it's a
simple matter of comparing the type to `minecraft:chest` and, if so,
setting the `LootTable` and `LootTableSeed` values for the NBT compound.

This commit simplifies the code to do precisely that: check for a chest,
set the relevant loot table and add a seed, and then return a new(?)
BlockInfo containing those. I'm unsure if returning the original
BlockInfo with the modified NBT would be sufficient also.

Conflicts:
	src/main/java/vazkii/quark/world/gen/structure/processor/BigDungeonChestProcessor.java