Update Shingles recipes to be compatible with Environmental
Environmental adds Terracotta Bricks and to be consistent with other bricks, they are crafted in a 2x2 and output 4. Unfortunately, Terracotta Shingles do the same. Since they are not bricks this changes them to be 1x2 since they have nothing to be consistent with and I personally think it makes a bit more sense for shingles and its a bit easier to craft since you need 2 for 2 instead of 4 for 4 while keeping the same ration.